load for plotting

+++ double/task1_funcs.c
@@ -1,6 +1,6 @@
 #include "task1_funcs.h"
 
-void print_matrix(float* a, int m, int n){
+void print_matrix(double* a, int m, int n){
 	for (int i=0; i<m; i++){
 		for (int j=0; j<n; j++){
 			printf("%f ", a[i*n + j]);
@@ -10,19 +10,19 @@ void print_matrix(float* a, int m, int n){
 }
 
 
-void init_matrix(float* matrix, int m, int n){
+void init_matrix(double* matrix, int m, int n){
 	for (int i=0; i<m; i++){
-		matrix[i*n + 0] = 0.98 * (float)((i+1)*(i+1)) / (float)(n*n); 
+		matrix[i*n + 0] = 0.98 * (double)((i+1)*(i+1)) / (double)(n*n); 
 	}
 	
 	for (int i=0; i<m; i++){	
 		for (int j=1; j<n; j++){	
-			matrix[i*n + j] = matrix[i*n + 0] * ( ((float)((m-j)*(m-j))) / ((float)(m*m)));	
+			matrix[i*n + j] = matrix[i*n + 0] * ( ((double)((m-j)*(m-j))) / ((double)(m*m)));	
 		}
 	}
 }
 
-void iterate(float* nextMatrix, float* previousMatrix, int m, int n){
+void iterate(double* nextMatrix, double* previousMatrix, int m, int n){
 	int mod_j_minus2;
 	int mod_j_minus1;
 	int mod_j_plus1;
@@ -41,23 +41,23 @@ void iterate(float* nextMatrix, float* previousMatrix, int m, int n){
 								previousMatrix[i*n + j] + 
 								(0.60*previousMatrix[i*n + mod_j_plus1]) +
 								(0.25*previousMatrix[i*n + mod_j_plus2]));
-			nextMatrix[i*n + j] /= (float)(5.0);
+			nextMatrix[i*n + j] /= (double)(5.0);
 		}
 	}
 }
 
-void calculate_avg_temp(float* matrix, int m, int n, float* thermometer){
+void calculate_avg_temp(double* matrix, int m, int n, double* thermometer){
 	for (int i=0; i<m; i++){	
 		for (int j=0; j<n; j++){	
 			thermometer[i] += matrix[i*n + j];
 		}
-		thermometer[i] /= n;
+		thermometer[i] /= (double)n;
 	}
 }
 
-float max_diff(float* a, float* b, int m, int n){
-	float diff;
-	float max_diff = 0;
+double max_diff(double* a, double* b, int m, int n){
+	double diff;
+	double max_diff = 0;
 
 	for (int i=0; i<m; i++){
 		for (int j=0; j<n; j++){
